+++ CHANGELOG.md
@@ -1,9 +1,22 @@
 # Change Log
 
+## 0.2.0
+## Changes
+### API
+- Trait `Clone` is no more required for keys
+- Renamed method `remove_item` to `remove`
+- Added method `remove_entry` which returns both key and priority
+- `push` operation returns old priority if had removed key now
+- Method `set_priority` returns `Result<(TPriority, ())>` with old priority instead of panicing on missing keys
+- Added Entry API to allow whole cycle `Find -> Read -> Update` with just one hashmap lookup.
+### Implementation
+- Now uses IndexMap from [indexmap](https://crates.io/crates/indexmap) crate internally
+
+
 ## 2020-02-25: 0.1.3
-## Changed
+## Changes
 - Removed unsafe implementations of Sync + Send because they are deduced by compiler
-- Make some optimizations which reduce timings by 50% but increase memory usage in worst case on 30%
+- Made some optimizations which reduce timings by 50% but increase memory usage in worst case on 30%
 - Added benchmarks
 
 ## 2019-11-24: 0.1.2
@@ -14,5 +27,5 @@
 ### Added
 - Now `KeyedPriorityQueue` implements `Default` trait
 
-### Changed
+### Changes
 - Some clippy fixes